requirements.txt
python-whois==0.6.4
requests==2.11.1